[Feat ✏️] PostCard Props 수정 및 반영
